--- src/components/explore-projects/ExploreProjectsStyles.js
@@ -1,31 +1,0 @@
-export const projectImagesStyles = {
-  objectFit: "cover",
-  height: "100%",
-  width: "100%",
-  position: "absolute",
-};
-export const onHoverClass = {
-  zIndex: "10",
-  transform: "scale(1.2)",
-  transition: "transform 0.4s ease-in-out",
-  cursor: "pointer",
-  opacity: "1",
-};
-export const onMouseLeaveStyle = {
-  zIndex: "10",
-  transform: "none",
-  boxShadow: "none",
-  opacity: "1",
-};
-export const noneHoveredProjectImagesStyle = {
-  zIndex: "1",
-  boxShadow: "none",
-};
-export const projectImagesContainer = {
-  margin: "0 auto",
-  marginBottom: "210px",
-  height: "1700px",
-  width: "810px",
-  overFlow: "visible",
-  padding: "50px 50px",
-};
